AB#1014 Move authorizationGroups pages
