+++ client/src/models/AuthorizationGroup.js
@@ -28,19 +28,6 @@ export default class AuthorizationGroup extends Model {
 
   static autocompleteQuery = "uuid name description"
 
-  static _resourceOverride = [
-    "admin",
-    utils.resourceize(this.resourceName)
-  ].join("/")
-
-  static pathFor(instance, query) {
-    return Model.pathFor(instance, query, this._resourceOverride)
-  }
-
-  static pathForNew(query) {
-    return Model.pathForNew(query, this._resourceOverride)
-  }
-
   static humanNameOfStatus(status) {
     return utils.sentenceCase(status)
   }
